@@ -78,9 +78,7 @@ def test_overlay(dfs_index, how):
     # construction of result
 
     def _read(name):
-        expected = read_file(
-            os.path.join(DATA, "polys", "df1_df2-{0}.geojson".format(name))
-        )
+        expected = read_file(os.path.join(DATA, "polys", f"df1_df2-{name}.geojson"))
         expected.geometry.array.crs = None
         for col in expected.columns[expected.dtypes == "int32"]:
             expected[col] = expected[col].astype("int64")
@@ -110,6 +108,22 @@ def test_overlay(dfs_index, how):
         result = result.reset_index(drop=True)
         expected = _read("difference-inverse")
         assert_geodataframe_equal(result, expected, check_column_type=False)
+
+
+@pytest.mark.parametrize("keep_geom_type", [True, False])
+def test_overlay_invalid_input(keep_geom_type):
+    invalid_polygon = Polygon([(0, 0), (1, 1), (1, 2), (1, 1), (0, 0)])
+    square = Polygon(((0, 0), (1, 0), (1, 1), (0, 1), (0, 0)))
+    df1 = geopandas.GeoDataFrame({"A": [1, 2]}, geometry=[invalid_polygon, square])
+    df2 = geopandas.GeoDataFrame({"B": [5, 6]}, geometry=[square, square])
+
+    df1_df2 = df1.overlay(df2, keep_geom_type=keep_geom_type)
+    # As overlay only supports inputs with uniform geometry types, `make_valid` of
+    # `invalid_polygon` should always result in POLYGON EMPTY, so the number of results
+    # will should always be 2.
+    assert len(df1_df2) == 2
+    if keep_geom_type:
+        assert df1_df2.geom_type.isin(["Polygon"]).all()
 
 
 @pytest.mark.filterwarnings("ignore:GeoSeries crs mismatch:UserWarning")
@@ -146,9 +160,7 @@ def test_overlay_nybb(how, nybb_filename):
         # read union one, further down below we take the appropriate subset
         expected = read_file(os.path.join(DATA, "nybb_qgis", "qgis-union.shp"))
     else:
-        expected = read_file(
-            os.path.join(DATA, "nybb_qgis", "qgis-{0}.shp".format(how))
-        )
+        expected = read_file(os.path.join(DATA, "nybb_qgis", f"qgis-{how}.shp"))
 
     # The result of QGIS for 'union' contains incorrect geometries:
     # 24 is a full original circle overlapping with unioned geometries, and
@@ -260,7 +272,7 @@ def test_overlay_overlap(how):
         raise pytest.skip()
 
     expected = read_file(
-        os.path.join(DATA, "overlap", "df1_df2_overlap-{0}.geojson".format(how))
+        os.path.join(DATA, "overlap", f"df1_df2_overlap-{how}.geojson")
     )
 
     if how == "union":
@@ -506,7 +518,7 @@ def test_overlay_strict(how, keep_geom_type, geom_types):
             os.path.join(
                 DATA,
                 "strict",
-                "{t}_{h}_{s}.geojson".format(t=geom_types, h=how, s=keep_geom_type),
+                f"{geom_types}_{how}_{keep_geom_type}.geojson",
             )
         )
 
@@ -813,6 +825,26 @@ def test_no_intersection():
     expected = GeoDataFrame(columns=["foo", "bar", "geometry"])
     result = overlay(gdf1, gdf2, how="intersection")
     assert_geodataframe_equal(result, expected, check_index_type=False)
+
+
+def test_zero_len():
+    # https://github.com/geopandas/geopandas/issues/3422
+    gdf1 = GeoDataFrame(
+        {
+            "geometry": [
+                Polygon([[0.0, 0.0], [2.0, 0.0], [2.0, 2.0], [0.0, 2.0], [0.0, 0.0]])
+            ]
+        },
+        crs=4326,
+    )
+    # overlay with empty geodataframe shouldn't throw
+    gdf2 = GeoDataFrame({"geometry": []}, crs=4326)
+    res = gdf1.overlay(gdf2, how="union")
+    assert_geodataframe_equal(res, gdf1)
+
+    gdf2 = GeoDataFrame(geometry=[], crs=4326)
+    res = gdf1.overlay(gdf2, how="union")
+    assert_geodataframe_equal(res, gdf1)
 
 
 class TestOverlayWikiExample:
